refactor(RouterInfo.jsx): 페이지 라우팅 기반 디렉토리 구조 변경
