+++ src/pages/collage/index.jsx
