+++ src/pages/feed/friends-list/index.jsx
